refactor: Rename old-scaffold to command-helpers/scaffold
